cwspo_repo/requirements.txt
accelerate>=0.34.0
bitsandbytes>=0.43.0; platform_system != "Darwin"
datasets>=2.20.0
numpy>=1.26.0
peft>=0.12.0
pydantic>=2.8.0
PyYAML>=6.0.1
rich>=13.7.1
sentencepiece>=0.2.0
torch>=2.3.0
tqdm>=4.66.4
transformers>=4.44.0
typer>=0.12.3
pytest>=8.2.0
